requirements (3).txt

flask==2.0.1
line-bot-sdk==2.1.0
